agent-finance-personal/pyproject.toml
[project]
name = "handler"
version = "0.1.0"
description = ""
authors = [
    {name = "Joseph Esteban",email = "esteban.programador@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "quart (>=0.20.0,<0.21.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
    "semantic-kernel (>=1.27.2,<2.0.0)",
    "quart-cors (>=0.8.0,<0.9.0)",
    "chainlit (>=2.5.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "handler", from = "src"}]

[tool.poetry.scripts]
start = "quart.cli:main --app src.handler:app"